refactor(Filters): broke filter controls into own component
