+++ src/pokemon-gens.json
@@ -1,0 +1,58 @@
+[
+  {
+    "id": "gen1",
+    "name": "Gen 1",
+    "range": {
+      "start": 1,
+      "end": 151
+    }
+  },
+  {
+    "id": "gen2",
+    "name": "Gen 2",
+    "range": {
+      "start": 152,
+      "end": 251
+    }
+  },
+  {
+    "id": "gen3",
+    "name": "Gen 3",
+    "range": {
+      "start": 252,
+      "end": 386
+    }
+  },
+  {
+    "id": "gen4",
+    "name": "Gen 4",
+    "range": {
+      "start": 387,
+      "end": 493
+    }
+  },
+  {
+    "id": "gen5",
+    "name": "Gen 5",
+    "range": {
+      "start": 494,
+      "end": 649
+    }
+  },
+  {
+    "id": "gen6",
+    "name": "Gen 6",
+    "range": {
+      "start": 650,
+      "end": 721
+    }
+  },
+  {
+    "id": "gen7",
+    "name": "Gen 7",
+    "range": {
+      "start": 722,
+      "end": 807
+    }
+  }
+]
